Restructure login page components and files
