+++ client/components/index.js
@@ -11,3 +11,4 @@ export { default as NavDesktop } from './nav-desktop';
 export { default as Container } from './container';
 export { default as InputField} from './input-field';
 export { default as Form } from './form';
+export { default as LoginPage } from './login-page';
